pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prompt-saver-mcp"
version = "2.0.0"
description = "MCP server for summarizing and saving conversation threads as prompts"
authors = [
    {name = "David Hou"},
]
dependencies = [
    "mcp[cli]>=1.12.0",
    "openai>=1.0.0",  # For LLM operations (OpenAI/Azure OpenAI)
    "anthropic>=0.25.0",  # For Anthropic Claude API
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",  # For YAML frontmatter in markdown files
    "typing-extensions>=4.0.0",
    "httpx>=0.24.0",  # For async HTTP requests
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
mongodb = [
    "motor>=3.3.0",  # Async MongoDB driver
    "pymongo>=4.6.0",  # MongoDB driver
]
embeddings = [
    "voyageai>=0.2.0",  # For vector embeddings
]
all = [
    "prompt-saver-mcp[mongodb,embeddings]",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
prompt-saver-mcp = "prompt_saver_mcp.server:main"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
